+++ queue_services/payment-reconciliations/requirements.txt
@@ -11,13 +11,13 @@ click==7.1.2
 itsdangerous==1.1.0
 jaeger-client==4.4.0
 jsonschema==3.2.0
-minio==7.0.1
+minio==7.0.2
 opentracing==2.4.0
 protobuf==3.14.0
 pycountry==20.7.3
 pyrsistent==0.17.3
 python-dotenv==0.15.0
-sentry-sdk==0.19.5
+sentry-sdk==0.20.2
 six==1.15.0
 threadloop==1.0.2
 thrift==0.13.0
