+++ queue_services/payment-reconciliations/requirements/prod.txt
@@ -8,4 +8,4 @@ pycountry
 Werkzeug<1
 minio
 jaeger-client
- attrs==19.1.0
+attrs==19.1.0
